@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "empyric"
-version = "0.2.6"
+version = "0.2.7"
 description = "A package for experiment automation"
 authors = ["Daniel Merthe <dmerthe@gmail.com>"]
 readme = "README.md"
@@ -15,9 +15,9 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.11"
 numpy = "1.24.3"
-scipy = "1.9.3"
+scipy = "^1.10"
 matplotlib = "3.6.2"
 pandas = "1.5.3"
 pykwalify = "1.8.0"
