@@ -1,9 +1,7 @@
 # Function/signal generators
-from typing import Union
-
+from empyric.types import ON, OFF, Toggle, String, Float
 from empyric.adapters import Socket
-from empyric.instruments import Instrument, setter, getter, measurer
-from empyric.types import ON, OFF, Toggle, Integer, String, Float, recast
+from empyric.instruments import Instrument, setter, getter
 
 
 class SiglentSDG1000(Instrument):
@@ -260,7 +258,7 @@ class SiglentSDG1000(Instrument):
 
     # Waveform high level
     @setter
-    def set_channel_1_high_level(self, high_level: Union[Float, String]):
+    def set_channel_1_high_level(self, high_level: Float):
         self._set_channel_n_basic_waveform(1, hlev=f"{high_level}")
 
     @getter
@@ -270,7 +268,7 @@ class SiglentSDG1000(Instrument):
         return float(high_level_str.replace("V", ""))
 
     @setter
-    def set_channel_2_high_level(self, high_level: Union[Float, String]):
+    def set_channel_2_high_level(self, high_level: Float):
         self._set_channel_n_basic_waveform(2, hlev=f"{high_level}")
 
     @getter
@@ -281,7 +279,7 @@ class SiglentSDG1000(Instrument):
 
     # Waveform low level
     @setter
-    def set_channel_1_low_level(self, low_level: Union[Float, String]):
+    def set_channel_1_low_level(self, low_level: Float):
         self._set_channel_n_basic_waveform(1, llev=f"{low_level}")
 
     @getter
@@ -291,7 +289,7 @@ class SiglentSDG1000(Instrument):
         return float(low_level_str.replace("V", ""))
 
     @setter
-    def set_channel_2_low_level(self, low_level: Union[Float, String]):
+    def set_channel_2_low_level(self, low_level: Float):
         self._set_channel_n_basic_waveform(2, llev=f"{low_level}")
 
     @getter
@@ -302,7 +300,7 @@ class SiglentSDG1000(Instrument):
 
     # Waveform frequency
     @setter
-    def set_channel_1_frequency(self, frequency: Union[Float, String]):
+    def set_channel_1_frequency(self, frequency: Float):
         self._set_channel_n_basic_waveform(1, frq=f"{frequency}")
 
     @getter
@@ -312,7 +310,7 @@ class SiglentSDG1000(Instrument):
         return float(freq_str.replace("HZ", ""))
 
     @setter
-    def set_channel_2_frequency(self, frequency: Union[Float, String]):
+    def set_channel_2_frequency(self, frequency: Float):
         self._set_channel_n_basic_waveform(2, frq=f"{frequency}")
 
     @getter
@@ -323,7 +321,7 @@ class SiglentSDG1000(Instrument):
 
     # Pulse width
     @setter
-    def set_channel_1_pulse_width(self, width: Union[Float, String]):
+    def set_channel_1_pulse_width(self, width: Float):
         self._set_channel_n_basic_waveform(1, width=f"{width}")
 
     @getter
@@ -333,7 +331,7 @@ class SiglentSDG1000(Instrument):
         return float(width_str.replace("S", ""))
 
     @setter
-    def set_channel_2_pulse_width(self, width: Union[Float, String]):
+    def set_channel_2_pulse_width(self, width: Float):
         self._set_channel_n_basic_waveform(2, width=f"{width}")
 
     @getter
@@ -344,7 +342,7 @@ class SiglentSDG1000(Instrument):
 
     # Pulse rise
     @setter
-    def set_channel_1_pulse_rise(self, rise: Union[Float, String]):
+    def set_channel_1_pulse_rise(self, rise: Float):
         self._set_channel_n_basic_waveform(1, rise=f"{rise}")
 
     @getter
@@ -354,7 +352,7 @@ class SiglentSDG1000(Instrument):
         return float(delay_str.replace("S", ""))
 
     @setter
-    def set_channel_2_pulse_rise(self, rise: Union[Float, String]):
+    def set_channel_2_pulse_rise(self, rise: Float):
         self._set_channel_n_basic_waveform(2, rise=f"{rise}")
 
     @getter
@@ -365,7 +363,7 @@ class SiglentSDG1000(Instrument):
 
     # Pulse fall
     @setter
-    def set_channel_1_pulse_fall(self, fall: Union[Float, String]):
+    def set_channel_1_pulse_fall(self, fall: Float):
         self._set_channel_n_basic_waveform(1, fall=f"{fall}")
 
     @getter
@@ -375,7 +373,7 @@ class SiglentSDG1000(Instrument):
         return float(delay_str.replace("S", ""))
 
     @setter
-    def set_channel_2_pulse_fall(self, fall: Union[Float, String]):
+    def set_channel_2_pulse_fall(self, fall: Float):
         self._set_channel_n_basic_waveform(2, fall=f"{fall}")
 
     @getter
@@ -386,7 +384,7 @@ class SiglentSDG1000(Instrument):
 
     # Pulse delay
     @setter
-    def set_channel_1_pulse_delay(self, delay: Union[Float, String]):
+    def set_channel_1_pulse_delay(self, delay: Float):
         self._set_channel_n_basic_waveform(1, dly=f"{delay}")
 
     @getter
@@ -396,7 +394,7 @@ class SiglentSDG1000(Instrument):
         return float(delay_str.replace("S", ""))
 
     @setter
-    def set_channel_2_pulse_delay(self, delay: Union[Float, String]):
+    def set_channel_2_pulse_delay(self, delay: Float):
         self._set_channel_n_basic_waveform(2, dly=f"{delay}")
 
     @getter
@@ -454,7 +452,7 @@ class SiglentSDG1000(Instrument):
 
     # Burst period
     @setter
-    def set_channel_1_burst_period(self, period: Union[Float, String]):
+    def set_channel_1_burst_period(self, period: Float):
         self._set_channel_n_burst_waveform(1, prd=f"{period}")
 
     @getter
@@ -464,7 +462,7 @@ class SiglentSDG1000(Instrument):
         return float(prd_str.replace("S", ""))
 
     @setter
-    def set_channel_2_burst_period(self, period: Union[Float, String]):
+    def set_channel_2_burst_period(self, period: Float):
         self._set_channel_n_burst_waveform(2, prd=f"{period}")
 
     @getter
@@ -492,7 +490,7 @@ class SiglentSDG1000(Instrument):
 
     # Burst trigger delay
     @setter
-    def set_channel_1_burst_trigger_delay(self, delay: Union[Float, String]):
+    def set_channel_1_burst_trigger_delay(self, delay: Float):
         self._set_channel_n_burst_waveform(1, dlay=f"{delay}")
 
     @getter
@@ -502,7 +500,7 @@ class SiglentSDG1000(Instrument):
         return float(dlay_str.replace("S", ""))
 
     @setter
-    def set_channel_2_burst_trigger_delay(self, delay: Union[Float, String]):
+    def set_channel_2_burst_trigger_delay(self, delay: Float):
         self._set_channel_n_burst_waveform(2, dlay=f"{delay}")
 
     @getter
