@@ -1,8 +1,10 @@
 # Tools for defining and running experiments
+import asyncio
 import collections
 import datetime
 import importlib
 import logging
+import numbers
 import os
 import pathlib
 import sys
@@ -10,6 +12,7 @@ import threading
 import time
 import tkinter as tk
 from tkinter.filedialog import askopenfilename
+from typing import Union
 
 import numpy as np
 import pandas as pd
@@ -23,16 +26,27 @@ from empyric import graphics as _graphics
 from empyric import instruments as _instruments
 from empyric import routines as _routines
 from empyric.adapters import AdapterError
-from empyric.tools import convert_time, Clock
+
+from empyric.tools import convert_time, Clock, logger
 from empyric.types import recast, Boolean, Toggle, Integer, Float, ON
 
-logger = logging.getLogger()
 
 class Experiment:
     """
     An iterable class which represents an experiment; iterates through any
-    assigned routines,and retrieves and stores the values of all experiment
-    variables.
+    assigned routines, and retrieves and stores the values of all experiment
+    variables. Each variable and routine is updated once per iteration, and each
+    iteration blocks until all variables and routines are updated.
+
+    The constructor take a `variables` argument in the form of a dictionary with the
+    format {..., name: variable, ...}, which contains all of the variables controlled
+    and monitored by the experiment. The optional `routines` argument is a dictionary
+    of the form {..., name: routine, ...} containing any routines to run within
+    the loop of the experiment. The optional `end` argument indicates when the
+    experiment should end (i.e. raise `StopIteration` on subsequent call to `__next__`)
+    , and can either be a number, a string of the form "[number] [time unit, e.g.
+    seconds, minutes or hours]" or "with routines" to end the experiment after the
+    last routine has ended.
     """
 
     # Possible statuses of an experiment
@@ -78,13 +92,19 @@ class Experiment:
     def terminated(self):
         return "Terminated" in self.status
 
-    def __init__(self, variables, routines=None, end=None):
+    def __init__(
+        self,
+        variables: dict,
+        routines: dict = None,
+        end: Union[numbers.Number, str, None] = None,
+    ):
         self.variables = variables
         # dict of the form {..., name: variable, ...}
 
         # Used to block evaluation of expressions
         # until their dependee variables are evaluated
-        self.eval_events = {name: threading.Event() for name in variables}
+        for variable in self.variables.values():
+            variable._eval_event = threading.Event()
 
         if routines:
             self.routines = routines
@@ -93,10 +113,15 @@ class Experiment:
             self.routines = {}
 
         if end:
-            if end.lower() == "with routines":
-                self.end = max([routine.end for routine in routines.values()])
-            else:
+            if type(end) is str:
+                if end.lower() == "with routines":
+                    self.end = max([routine.end for routine in routines.values()])
+                else:
+                    self.end = convert_time(end)
+            elif isinstance(end, numbers.Number):
                 self.end = end
+            else:
+                raise ValueError("invalid value for end keyword argument")
         else:
             self.end = np.inf
 
@@ -105,10 +130,12 @@ class Experiment:
 
         self.timestamp = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
 
+        self.state = pd.Series(
+            name=None,
+            data={**{"Time": None}, **{name: None for name in self.variables}},
+            dtype=object,
+        )
         self.data = pd.DataFrame(columns=["Time"] + list(variables.keys()))
-
-        self.state = pd.Series({column: None for column in self.data.columns})
-        self.state["Time"] = 0
 
         self._status = Experiment.READY
         self.status_locked = True
@@ -117,14 +144,17 @@ class Experiment:
         self.saved = []  # list of saved data entries
 
     def __next__(self):
+
         # Start the clock on first call
         if self.state.name is None:  # first step of the experiment
             self.start()
-            self.status = Experiment.RUNNING + ": initializing..."
+            self.status = Experiment.RUNNING
 
         # Update time
         self.state["Time"] = self.clock.time
         self.state.name = datetime.datetime.now()
+
+        logger.info(f'Iterating experiment (t = {self.state["Time"]} s)')
 
         if self.stopped:
             return self.state
@@ -149,8 +179,8 @@ class Experiment:
 
         # Get all variable values if experiment is running or holding
         if self.running or self.holding:
-            for event in self.eval_events.values():
-                event.clear()
+            for variable in self.variables.values():
+                variable._eval_event.clear()
 
             # Run each measure / get operation in its own thread
             threads = {}
@@ -189,18 +219,26 @@ class Experiment:
 
         try:
             if isinstance(self.variables[name], _variables.Expression):
-                for dependee in self.variables[name].definitions:
-                    event = self.eval_events[dependee]
-                    event.wait()  # wait for dependee to be evaluated
+                for symbol, dependee in self.variables[name].definitions.items():
+                    if hasattr(dependee, "_eval_event"):
+
+                        logger.debug(
+                            f'Expression {name} is waiting for {symbol} to be evaluated'
+                        )
+
+                        dependee._eval_event.wait()
 
             try:
                 value = self.variables[name].value
-            except AdapterError:
+                logger.info(f'{name} evaluated to {value}')
+            except AdapterError as adapter_error:
                 value = None
-            except ValueError:
+                logger.warning(f'Unable to evaluate {name}: {adapter_error}')
+            except ValueError as value_error:
                 value = None
+                logger.warning(f'Unable to evaluate {name}: {value_error}')
 
-            self.eval_events[name].set()
+            self.variables[name]._eval_event.set()
             # unblock threads evaluating dependents
 
             if np.size(value) > 1:  # store array data as CSV files
@@ -218,7 +256,7 @@ class Experiment:
                     self.state[name] = None
             else:
                 self.state[name] = value
-        except BaseException as err:
+        except Exception as err:
             self.terminate()
             raise err
 
@@ -226,8 +264,11 @@ class Experiment:
         """Update a routine according to the current state"""
 
         try:
-            self.routines[name].update(self.state)
-        except BaseException as err:
+            try:
+                self.routines[name].update(self.state)
+            except AdapterError as adapter_error:
+                logger.warning(f'Unable to update routine {name}: {adapter_error}')
+        except Exception as err:
             self.terminate()
             raise err
 
@@ -247,6 +288,8 @@ class Experiment:
 
         if directory:
             path = os.path.join(directory, path)
+
+        logger.info(f'Saving experiment data to {path}')
 
         if not os.path.exists(path):
             # if this is a new file, write column headers
@@ -273,6 +316,8 @@ class Experiment:
         """
         self.clock.start()
 
+        logger.info('Starting experiment')
+
         self.status_locked = False
         self.status = Experiment.RUNNING
         self.status_locked = True
@@ -284,6 +329,8 @@ class Experiment:
         :return: None
         """
         self.clock.stop()
+
+        logger.info(f'Holding experiment ({reason})')
 
         self.status_locked = False
         self.status = Experiment.HOLDING
@@ -299,6 +346,8 @@ class Experiment:
         """
         self.clock.stop()
 
+        logger.info(f'Stopping experiment ({reason})')
+
         self.status_locked = False
         self.status = Experiment.STOPPED
         if reason:
@@ -312,6 +361,9 @@ class Experiment:
 
         :return: None
         """
+
+        logger.info(f'Terminating experiment ({reason})')
+
         self.stop()
         self.save()
 
@@ -329,15 +381,96 @@ class Experiment:
         return "Experiment"
 
 
+class AsyncExperiment(Experiment):
+    """
+    Asynchronous version of Experiment
+
+    Each variable and routine is updated as quickly as possible independent of the
+    experiment iteration. Every time a variable is updated, the corresponding entry in
+    `state` is also updated.
+    """
+
+    def __init__(
+        self,
+        variables: dict,
+        routines: dict = None,
+        end: Union[numbers.Number, str, None] = None,
+    ):
+        super().__init__(variables, routines, end)
+        self.loop = None
+
+    def __next__(self):
+
+        # Start the clock and loop on first call
+        if self.state.name is None:  # first step of the experiment
+            self.start()
+            self.status = Experiment.RUNNING
+
+        logger.info(f'Iterating experiment (t = {self.state["Time"]} s)')
+
+        # End the experiment, if the duration of the experiment has passed
+        if self.clock.time > self.end:
+            self.terminate()
+
+        if (self.running or self.holding) and self.state.name is not None:
+            # Append new state to experiment data set
+            self.data.loc[self.state.name] = self.state
+
+        elif self.terminated:
+            raise StopIteration
+
+        return self.state
+
+    def start(self):
+
+        super().start()
+
+        self.loop = asyncio.get_running_loop()
+
+        for name in self.variables:
+            self.loop.create_task(self._update_variable(name))
+
+        for name in self.routines:
+            self.loop.create_task(self._update_routine(name))
+
+    async def _update_variable(self, name):
+        """Update named variable"""
+        while not self.terminated:
+            if self.running or self.holding:
+
+                await asyncio.to_thread(Experiment._update_variable, self, name)
+
+                # Update time
+                self.state["Time"] = self.clock.time
+                self.state.name = datetime.datetime.now()
+
+            elif self.stopped:
+                await asyncio.sleep(0)  # give other updating tasks a chance to run
+
+    async def _update_routine(self, name):
+        """Update named routine"""
+        while not self.terminated:
+            if self.running:
+
+                # Update time
+                self.state["Time"] = self.clock.time
+                self.state.name = datetime.datetime.now()
+
+                await asyncio.to_thread(Experiment._update_routine, self, name)
+
+            elif self.holding or self.stopped:
+                await asyncio.sleep(0)  # give other updating tasks a chance to run
+
+
 class Alarm:
     """
     Triggers if a condition among variables is met and indicates the response
     protocol
     """
 
-    def __init__(self, condition, variables, protocol=None):
+    def __init__(self, condition, definitions, protocol=None):
         self.trigger_variable = _variables.Expression(
-            expression=condition, definitions=variables
+            expression=condition, definitions=definitions
         )
 
         if protocol:
@@ -347,7 +480,7 @@ class Alarm:
 
     @property
     def triggered(self):
-        return self.trigger_variable.value is True
+        return bool(self.trigger_variable.value)
 
     def __repr__(self):
         return "Alarm"
@@ -367,6 +500,8 @@ class Manager:
         a YAML file
         """
 
+        logger.info(f"Initializing experiment manager...")
+
         if runcard:
             self.runcard = runcard
         else:
@@ -383,6 +518,9 @@ class Manager:
                 raise ValueError("a valid runcard was not selected!")
 
         if isinstance(self.runcard, str):
+
+            logger.info(f'Parsing runcard from file ({self.runcard})...')
+
             if os.path.exists(self.runcard):
                 dirname = os.path.dirname(self.runcard)
                 if dirname != "":
@@ -395,7 +533,9 @@ class Manager:
             else:
                 raise FileNotFoundError(f'invalid runcard path "{self.runcard}"')
         elif isinstance(self.runcard, dict):
-            pass
+
+            logger.info(f'Parsing runcard from dictionary...')
+
         else:
             raise TypeError(
                 f"runcard given to Manager must be a path to a YAML file "
@@ -410,6 +550,8 @@ class Manager:
         self.alarms = converted_runcard["Alarms"]
         self.plotter = converted_runcard.get("Plotter", None)
 
+        logger.info('Building experiment from runcard...')
+
         self.experiment = converted_runcard["Experiment"]
 
         # Unpack settings
@@ -417,9 +559,6 @@ class Manager:
         self.step_interval = convert_time(self.settings.get("step interval", 0.1))
         self.save_interval = convert_time(self.settings.get("save interval", 60))
         self.plot_interval = convert_time(self.settings.get("plot interval", 0.1))
-        self.end = convert_time(self.settings.get("end", np.inf))
-
-        self.experiment.end = self.end
 
         self.last_save = 0
 
@@ -454,10 +593,12 @@ class Manager:
             yaml.dump(self.runcard, runcard_file)
 
         # Run experiment loop in separate thread
-        experiment_thread = threading.Thread(target=self._run)
+        logger.info(f'Starting {experiment_name}')
+        experiment_thread = threading.Thread(target=asyncio.run, args=(self._run(),))
         experiment_thread.start()
 
         # Set up the GUI for user interaction
+        logger.info('Launching GUI')
         self.gui = _graphics.ExperimentGUI(
             self.experiment,
             alarms=self.alarms,
@@ -471,8 +612,10 @@ class Manager:
         self.gui.run()
 
         experiment_thread.join()
+        logger.info('Experiment complete; cleaning up...')
 
         # Disconnect instruments
+        logger.info('Disconnecting from instruments')
         for instrument in self.instruments.values():
             instrument.disconnect()
 
@@ -480,10 +623,12 @@ class Manager:
 
         # Cancel follow-up if experiment terminated by user
         if self.experiment.terminated and "user" in self.experiment.status:
+            logger.info('Cancelling follow-up experiment')
             self.followup = None
 
         # Execute the follow-up experiment if there is one
         if self.followup:
+            logger.info('Running follow-up experiment')
             if "yaml" in self.followup:
                 self.__init__(self.followup)
                 self.run(directory=directory)
@@ -491,10 +636,17 @@ class Manager:
                 self.__init__(self.runcard)
                 self.run(directory=directory)
 
-    def _run(self):
+    async def _run(self):
         # Run in a separate thread
 
         for state in self.experiment:
+
+            logger.info(
+                'state =' + '.,'.join(
+                    [f'{key}: {value}' for key, value in state.items()]
+                )
+            )
+
             step_start = time.time()
 
             # Save experimental data periodically
@@ -567,8 +719,7 @@ class Manager:
 
             remaining_time = self.step_interval - (step_end - step_start)
 
-            if remaining_time > 0:
-                time.sleep(remaining_time)
+            await asyncio.sleep(max([remaining_time, 0.0]))
 
     def __repr__(self):
         return "Manager"
@@ -579,6 +730,9 @@ class RuncardError(BaseException):
 
 
 def validate_runcard(runcard):
+
+    logger.info('Validating runcard')
+
     is_dict = isinstance(runcard, dict)
     is_ordereddict = isinstance(runcard, collections.OrderedDict)
 
@@ -591,13 +745,14 @@ def validate_runcard(runcard):
         with open(runcard_path, "w") as runcard_file:
             yaml.dump(runcard, runcard_file)
 
-        validate_runcard(runcard_path)
+        try:
+            validate_runcard(runcard_path)
+        finally:
+            # Wait until temporary file has write access, then delete
+            while not os.access(runcard_path, os.W_OK):
+                time.sleep(0.1)
 
-        # Wait until temporary file has write access, then delete
-        while not os.access(runcard_path, os.W_OK):
-            time.sleep(0.1)
-
-        os.remove(runcard_path)
+            os.remove(runcard_path)
 
         return True
 
@@ -637,6 +792,8 @@ def convert_runcard(runcard):
     * The Plots section is converted into a corresponding ``Plotter`` instance.
     """
 
+    logger.info('Building experiment from runcard')
+
     # if runcard argument is a path to a YAML file, load into a dictionary
     if type(runcard) == str:
         yaml = YAML()
@@ -668,6 +825,9 @@ def convert_runcard(runcard):
 
     instruments = {}
     for name, specs in runcard.get("Instruments", {}).items():
+
+        logger.info(f'Initializing instrument {name}')
+
         specs = specs.copy()
         _type = specs.pop("type")
         address = specs.get("address", None)
@@ -697,12 +857,19 @@ def convert_runcard(runcard):
     # Variables section
     variables = {}
     for name, specs in runcard["Variables"].items():
+
+        logger.info(f'Initializing variable {name}')
+
         if "meter" in specs:
             instrument = converted_runcard["Instruments"][specs["instrument"]]
             gate = specs.get("gate", None)
             gate = variables[gate] if gate else None
             variables[name] = _variables.Meter(
-                meter=specs["meter"], instrument=instrument, gate=gate
+                meter=specs["meter"],
+                instrument=instrument,
+                gate=gate,
+                multiplier=specs.get("multiplier", 1),
+                offset=specs.get("offset", 0),
             )
         elif "knob" in specs:
             instrument = converted_runcard["Instruments"][specs["instrument"]]
@@ -711,21 +878,23 @@ def convert_runcard(runcard):
                 instrument=instrument,
                 lower_limit=specs.get("lower limit", None),
                 upper_limit=specs.get("upper limit", None),
+                multiplier=specs.get("multiplier", 1),
+                offset=specs.get("offset", 0),
             )
         elif "expression" in specs:
             expression = specs["expression"]
-            definitions = {}
+            definitions = specs.get("definitions", {}).copy()
 
-            for symbol, var_name in specs["definitions"].items():
-                expression = expression.replace(symbol, var_name)
-
-                try:
-                    definitions[var_name] = variables[var_name]
-                except KeyError as undefined:
+            for variable in definitions.values():
+                if variable not in variables:
                     raise KeyError(
-                        f"variable {undefined} is not defined for expression "
-                        f"'{name}'"
+                        f"variable {variable} specified for expression {name} "
+                        f"is not in Variables!"
                     )
+
+            definitions = {
+                symbol: variables[name] for symbol, name in definitions.items()
+            }
 
             variables[name] = _variables.Expression(
                 expression=expression, definitions=definitions
@@ -737,7 +906,12 @@ def convert_runcard(runcard):
             settable = specs.get("settable", False)
 
             variables[name] = _variables.Remote(
-                server=server, alias=alias, protocol=protocol, settable=settable
+                server=server,
+                alias=alias,
+                protocol=protocol,
+                settable=settable,
+                multiplier=specs.get("multiplier", 1),
+                offset=specs.get("offset", 0),
             )
 
         elif "parameter" in specs:
@@ -754,6 +928,9 @@ def convert_runcard(runcard):
     routines = {}
     if "Routines" in runcard:
         for name, specs in runcard["Routines"].items():
+
+            logger.info(f'Initializing routine {name}')
+
             specs = specs.copy()  # avoids modifying the runcard
             _type = specs.pop("type")
 
@@ -773,57 +950,52 @@ def convert_runcard(runcard):
 
             routines[name] = available_routines[_type](**specs)
 
-    converted_runcard["Experiment"] = Experiment(variables, routines=routines)
+    # Create the experiment
+    async_experiment = False
+    if "Settings" in runcard:
+        if runcard["Settings"].get("async", False):
+            async_experiment = True
+
+    if async_experiment:
+
+        logger.info('Initializing synchronous experiment')
+
+        converted_runcard["Experiment"] = AsyncExperiment(
+            variables, routines=routines, end=runcard["Settings"].get("end", None)
+        )
+    else:
+
+        logger.info('Initializing asynchronous experiment')
+
+        converted_runcard["Experiment"] = Experiment(
+            variables, routines=routines, end=runcard["Settings"].get("end", None)
+        )
 
     # Alarms section
     alarms = {}
     if "Alarms" in runcard:
         for name, specs in runcard["Alarms"].items():
-            alarm_variables = specs.copy().get("variables", {})
-            condition = specs.copy()["condition"]
 
-            for variable in specs.get("variables", {}):
+            logger.info(f'Initializing alarm {name}')
+
+            condition = specs.copy()["condition"]
+            definitions = specs.copy().get("definitions", {})
+
+            for variable in definitions.values():
                 if variable not in variables:
                     raise KeyError(
                         f"variable {variable} specified for alarm {name} "
                         f"is not in Variables!"
                     )
 
-            alphabet = "abcdefghijklmnopqrstuvwxyz"
-            for var_name, variable in variables.items():
-                # Variables can be called by name in the condition
-                if var_name in condition:
-                    temp_name = "".join(
-                        [
-                            alphabet[  # pylint: disable=invalid-sequence-index
-                                np.random.randint(0, len(alphabet))
-                            ]
-                            for i in range(3)
-                        ]
-                    )
-                    while temp_name in alarm_variables:
-                        # make sure temp_name is not repeated
-                        temp_name = "".join(
-                            [
-                                alphabet[  # pylint: disable=invalid-sequence-index
-                                    np.random.randint(0, len(alphabet))
-                                ]
-                                for i in range(3)
-                            ]
-                        )
-
-                    alarm_variables.update({temp_name: variable})
-                    condition = condition.replace(var_name, temp_name)
-
-                # Otherwise, they are defined in the alarm_variables
-                for symbol, alarm_variable_name in alarm_variables.items():
-                    if alarm_variable_name == var_name:
-                        alarm_variables[symbol] = variable
+            definitions = {
+                symbol: variables[name] for symbol, name in definitions.items()
+            }
 
             alarms.update(
                 {
                     name: Alarm(
-                        condition, alarm_variables, protocol=specs.get("protocol", None)
+                        condition, definitions, protocol=specs.get("protocol", None)
                     )
                 }
             )
@@ -832,6 +1004,9 @@ def convert_runcard(runcard):
 
     # Plots section
     if "Plots" in runcard:
+
+        logger.info('Initializing plotter')
+
         converted_runcard["Plotter"] = _graphics.Plotter(
             converted_runcard["Experiment"].data, settings=runcard["Plots"]
         )
