@@ -2,8 +2,26 @@ import time
 import select
 import socket
 import numbers
+import logging
 import numpy as np
+import logging
 
+# Set up logging
+logger = logging.getLogger("empyric")
+
+logger.setLevel(logging.WARNING)
+
+log_stream_handler = logging.StreamHandler()
+log_stream_handler.setLevel(logging.WARNING)
+log_stream_handler.setFormatter(
+    logging.Formatter(
+        '%(asctime)s: %(message)s'
+    )
+)
+
+logger.addHandler(log_stream_handler)
+
+# TODO add logging.FileHandler to drop log into file in working directory
 
 # Tools for time-keeping
 def convert_time(time_value):
@@ -141,7 +159,7 @@ def autobind_socket(_socket):
 
 
 def read_from_socket(
-    _socket, nbytes=None, termination="\r", timeout=1, decode=True, chunk_size=4096
+    _socket, nbytes=None, termination="\r", timeout=None, decode=True, chunk_size=4096
 ):
     """
     Read from a socket, with some effort taken to get the whole message.
@@ -155,13 +173,16 @@ def read_from_socket(
                         otherwise.
     :param timeout: (numbers.Number) communication timeout in seconds;
                     used for both the `select.select` and `_socket.recv`
-                    functions.
+                    functions; defaults to existing timeout of _socket.
     :param decode: (bool) whether to return decoded string (True) or raw bytes
                    message (False); defaults to True.
     :param chunk_size: (int) number of bytes to read on each call to recv method.
     """
 
     default_timeout = _socket.gettimeout()  # save default timeout
+
+    if timeout is None:
+        timeout = default_timeout
 
     _socket.settimeout(timeout)
 
@@ -209,11 +230,6 @@ def read_from_socket(
             raise ConnectionError("an exception has been raised for the socket")
 
         try:
-            # Check again that socket is readable
-            readable = select.select([_socket], [], [], timeout)[0]
-
-            if not readable:
-                break
 
             if remaining_bytes < chunk_size:
                 part = _socket.recv(remaining_bytes)
@@ -221,7 +237,10 @@ def read_from_socket(
                 part = _socket.recv(chunk_size)
 
         except ConnectionResetError as err:
-            print(f"Warning: {err}")
+            print(
+                f"Warning: while reading from socket at {_socket.getsockname()}, "
+                f"got error: {err}"
+            )
             break
         except socket.timeout:
             pass
@@ -243,7 +262,7 @@ def read_from_socket(
         return message
 
 
-def write_to_socket(_socket, message, termination="\r", timeout=1):
+def write_to_socket(_socket, message, termination="\r", timeout=None):
     """
     Write a message to a socket, with care taken to get the whole message
     transmitted.
@@ -251,24 +270,19 @@ def write_to_socket(_socket, message, termination="\r", timeout=1):
     :param _socket: (socket.Socket) socket to write to.
     :param message: (str/bytes) message to send.
     :param termination: (str/bytes) expected message termination character(s).
-    :param timeout: (numbers.Number) timeout for `select.select` call.
+    :param timeout: (numbers.Number) timeout for `select.select` call; defaults to
+    existing timeout of _socket.
     """
 
-    # Block until the socket is writeable or until timeout
-    if timeout:
-        writeable = _socket in select.select([], [_socket], [], timeout)[1]
-    else:
-        writeable = _socket in select.select([], [_socket], [])[1]
-
-    if not writeable:
-        return 0
+    if timeout is None:
+        timeout = _socket.gettimeout()
 
     if isinstance(message, str):
         bytes_message = (message + termination).encode()
     elif isinstance(message, bytes):
         bytes_message = message + termination.encode()
     else:
-        raise ValueError('message argument must be either string or bytes')
+        raise ValueError("message argument must be either string or bytes")
 
     msg_len = len(bytes_message)
 
@@ -278,6 +292,15 @@ def write_to_socket(_socket, message, termination="\r", timeout=1):
     total_sent = 0
 
     while total_sent < msg_len and failures < max_failures:
+        # Block until the socket is writeable or until timeout
+        if timeout:
+            writeable = _socket in select.select([], [_socket], [], timeout)[1]
+        else:
+            writeable = _socket in select.select([], [_socket], [])[1]
+
+        if not writeable:
+            break
+
         sent = _socket.send(bytes_message[total_sent:])
 
         if sent == 0:
@@ -286,6 +309,8 @@ def write_to_socket(_socket, message, termination="\r", timeout=1):
         total_sent = total_sent + sent
 
     if total_sent < msg_len:
-        raise ConnectionError(f"Socket connection to {_socket.getsockname()} is broken!")
+        raise ConnectionError(
+            f"Socket connection to {_socket.getsockname()} is broken!"
+        )
 
     return total_sent
