@@ -5,12 +5,16 @@ import socket
 import time
 import typing
 from functools import wraps
-import numpy as np  # functions used in Expression's eval call
+
+import dill
+import numpy as np  # used in Expression's eval call
+
 from empyric.collection.instrument import Instrument
 
-from empyric import instruments, types
-from empyric.tools import write_to_socket, read_from_socket
-from empyric.types import *
+from empyric.instruments import ModbusClient
+from empyric.tools import write_to_socket, read_from_socket, logger
+from empyric.types import supported as supported_types, recast
+from empyric.types import Type, Boolean, Float, Integer, Toggle, ON, Array
 
 
 class Variable:
@@ -19,27 +23,33 @@ class Variable:
     experiment.
     """
 
-    _type = None  #: the data type of the variable
+    _type = None  # the data type of the variable
 
     #: time since the epoch of last evaluation in seconds, being equal to the
     #: result of `time.time()` being called upon the most recent evaluation of
     #: the `value` property
     last_evaluation = None
 
-    _settable = False  #: whether the variable can be set by the user
-    _value = None  #: last known value of the variable
+    _settable = False  # whether the variable can be set by the user
+    _value = None  # last known value of the variable
 
-    _hidden = False  #: used by GUIs
+    _hidden = False  # used by GUIs
+
+    @property
+    def type(self):
+        """Data type of the variable"""
+        return self._type
 
     @property
     def settable(self):
+        """Whether values can be assigned to the variable by the user"""
         return self._settable
 
     @property
     def value(self):
         """The value of the variable"""
         # overwritten by child classes
-        pass
+        return
 
     @value.setter
     def value(self, value):
@@ -65,7 +75,7 @@ class Variable:
 
                 recasted_value = recast(value)
 
-                for _type in types.supported.values():
+                for _type in supported_types.values():
                     if isinstance(recasted_value, _type):
                         self._type = _type
                         setter(self, recasted_value)
@@ -93,7 +103,7 @@ class Variable:
 
                 recasted_value = recast(value)
 
-                for _type in types.supported.values():
+                for _type in supported_types.values():
                     if isinstance(recasted_value, _type):
                         self._type = _type
                         self._value = recasted_value
@@ -101,6 +111,33 @@ class Variable:
             return self._value
 
         return wrapped_getter
+
+    def __mul__(self, other):
+        if isinstance(other, Variable):
+            return self._value * other._value
+        else:
+            return self._value * other
+
+    def __add__(self, other):
+        if isinstance(other, Variable):
+            return self._value + other._value
+        else:
+            return self._value + other
+
+    def __sub__(self, other):
+        if isinstance(other, Variable):
+            return self._value - other._value
+        else:
+            return self._value - other
+
+    def __bool__(self):
+        return bool(self._value)
+
+    def __eq__(self, other):
+        if isinstance(other, Variable):
+            return self._value == other._value
+        else:
+            return self._value == other
 
 
 class Knob(Variable):
@@ -112,6 +149,16 @@ class Knob(Variable):
     associated with.
 
     The `knob` argument is the label of the knob on the instrument.
+
+    The optional `lower_limit` and `upper_limit` keyword arguments set the lower and
+    upper limits for the knob. If an attempt it made to set a value above/below the
+    upper/lower limit, it is set to the upper/lower limit value.
+
+    The optional `multiplier` and `offset` keyword arguments provide a means to affect
+    a linear transformation of the raw knob value. Readings from the instrument will
+    be multiplied by the `multiplier` and then increased by the `offset`. Set commands
+    to the instrument will take the knob value, subtract the `offset` and divide by the
+    `multiplier`.
     """
 
     _settable = True  #:
@@ -120,13 +167,17 @@ class Knob(Variable):
         self,
         instrument: Instrument,
         knob: str,
-        lower_limit: numbers.Number = None,
-        upper_limit: numbers.Number = None,
+        lower_limit: typing.Union[float, int] = None,
+        upper_limit: typing.Union[float, int] = None,
+        multiplier: typing.Union[float, int] = 1,
+        offset: typing.Union[float, int] = 0,
     ):
         self.instrument = instrument
         self.knob = knob  # name of the knob on instrument
         self.lower_limit = lower_limit
         self.upper_limit = upper_limit
+        self.multiplier = multiplier
+        self.offset = offset
 
         # infer type from type hint of first argument of set method
         set_method = getattr(instrument, "set_" + knob.replace(" ", "_"))
@@ -136,6 +187,14 @@ class Knob(Variable):
         if type_hints:
             arg_hints = list(type_hints)
             self._type = type_hints[arg_hints[0]]
+        else:
+
+            logger.warning(
+                f"Unable to determine data dtype of {knob} on {instrument}; "
+                "assuming 64-bit float"
+            )
+
+            self._type = np.float64
 
         self._value = None
 
@@ -147,6 +206,10 @@ class Knob(Variable):
         """
 
         self._value = self.instrument.get(self.knob)
+
+        if isinstance(self._value, numbers.Number):
+            self._value = self.multiplier * self._value + self.offset
+
         self.last_evaluation = time.time()
 
         return self._value
@@ -158,14 +221,36 @@ class Knob(Variable):
         Set an instrument knob to value
         """
 
-        if self.upper_limit and value > self.upper_limit:
-            self.instrument.set(self.knob, self.upper_limit)
-        elif self.lower_limit and value < self.lower_limit:
-            self.instrument.set(self.knob, self.lower_limit)
-        else:
-            self.instrument.set(self.knob, value)
+        try:
+            if self.upper_limit and value > self.upper_limit:
+                self.instrument.set(
+                    self.knob, (self.upper_limit - self.offset) / self.multiplier
+                )
+            elif self.lower_limit and value < self.lower_limit:
+                self.instrument.set(
+                    self.knob, (self.lower_limit - self.offset) / self.multiplier
+                )
+            else:
+                if isinstance(value, numbers.Number):
+                    self.instrument.set(
+                        self.knob, (value - self.offset) / self.multiplier
+                    )
+                else:
+                    self.instrument.set(self.knob, value)
 
-        self._value = self.instrument.__getattribute__(self.knob.replace(" ", "_"))
+        except TypeError as type_error:
+
+            logger.warning(str(type_error))
+
+        self._value = self.instrument.__getattribute__(
+            self.knob.replace(" ", "_")
+        )
+
+        if isinstance(self._value, numbers.Number):
+            self._value = self.multiplier * self._value + self.offset
+
+    def __str__(self):
+        return f"Knob({self.value})"
 
 
 class Meter(Variable):
@@ -186,13 +271,26 @@ class Meter(Variable):
     it should be a variable of integer, boolean or toggle type. When the gate
     variable evaluates to 1/True/On, the meter can be measured. Otherwise,
     attempts to measure the meter will have no effect (`None` is returned).
+
+    The optional `multiplier` and `offset` keyword arguments provide a means to affect
+    a linear transformation of the raw meter value. Readings from the instrument will
+    be multiplied by the `multiplier` and then increased by the `offset`.
     """
 
     _settable = False  #:
 
-    def __init__(self, instrument: Instrument, meter: str, gate=None):
+    def __init__(
+        self,
+        instrument: Instrument,
+        meter: str,
+        gate=None,
+        multiplier: typing.Union[float, int] = 1,
+        offset: typing.Union[float, int] = 0,
+    ):
         self.instrument = instrument
         self.meter = meter
+        self.multiplier = multiplier
+        self.offset = offset
 
         if gate and isinstance(gate, Variable):
             self.gate = gate
@@ -201,7 +299,7 @@ class Meter(Variable):
 
         self._type = typing.get_type_hints(
             getattr(instrument, "measure_" + meter.replace(" ", "_"))
-        ).get("return", None)
+        ).get("return", np.float64)
 
         self._value = None
 
@@ -216,9 +314,16 @@ class Meter(Variable):
             return None
 
         self._value = self.instrument.measure(self.meter)
+
+        if isinstance(self._value, numbers.Number):
+            self._value = self.multiplier * self._value + self.offset
+
         self.last_evaluation = time.time()
 
         return self._value
+
+    def __str__(self):
+        return f"Meter({self.value})"
 
 
 class Expression(Variable):
@@ -239,6 +344,8 @@ class Expression(Variable):
 
     _settable = False  #:
 
+    # shorthand terms for common functions
+    # TODO consolidate these functions in a separate module
     _functions = {
         "sqrt(": "np.sqrt(",
         "exp(": "np.exp(",
@@ -253,6 +360,11 @@ class Expression(Variable):
         "diff(": "np.diff(",
         "max(": "np.nanmax(",
         "min(": "np.nanmin(",
+        "fft(": "self.fft(",
+        "ifft(": "self.ifft(",
+        "carrier(": "self.carrier(",
+        "ampl(": "self.ampl(",
+        "demod(": "self.demod(",
     }
 
     def __init__(self, expression: str, definitions: dict = None):
@@ -268,31 +380,178 @@ class Expression(Variable):
 
         expression = self.expression
 
+        logger.debug(f'Evaluating expression {expression}')
+
         # carets represent exponents
         expression = expression.replace("^", "**")
 
-        for symbol, variable in self.definitions.items():
-            expression = expression.replace(symbol, f"({variable._value})")
+        variables = {
+            symbol: variable._value for symbol, variable in self.definitions.items()
+        }
 
         for shorthand, longhand in self._functions.items():
             if shorthand in expression:
                 expression = expression.replace(shorthand, longhand)
 
         try:
-            if "None" not in expression and "nan" not in expression:
-                self._value = eval(expression)
-            else:
-                self._value = None
-        except BaseException as err:
-            print(
-                f"Unable to evaluate expression {self.expression} due to " f"error: ",
-                err,
+            all_values = np.concatenate(
+                [np.atleast_1d(val).flatten() for val in variables.values()]
             )
+
+            no_nones = None not in all_values
+            no_nans = np.nan not in all_values
+            no_infs = (np.inf not in all_values) and (-np.inf not in all_values)
+
+            valid_values = no_nones and no_nans and no_infs
+
+            if valid_values:
+                self._value = eval(expression, {**globals(), **variables}, locals())
+            else:
+
+                log_str = (
+                    f"Dependencies for {expression} contain invalid values: "
+                )
+
+                for name, value in variables.items():
+
+                    if value is None:
+                        log_str += f"\n{name} = None"
+                    elif np.isnan(value):
+                        log_str += f"\n{name} = NaN"
+                    elif np.isinf(value):
+                        log_str += f"\n{name} = +/-Inf"
+
+                logger.debug(log_str)
+
+                self._value = None
+
+        except Exception as err:
+
+            logger.warning(
+                f"Unable to evaluate expression {self.expression} due to error: {err}"
+            )
+
             self._value = None
 
         self.last_evaluation = time.time()
 
+        logger.debug(f'Expression {self.expression} evaluated to {self._value}')
+
         return self._value
+
+    def __str__(self):
+        if len(str(self.value)) < 100:  # first call to value evaluates expression
+            return f"Expression({self.expression} = {self._value})"
+        elif isinstance(self._value, Array):
+            return f"Expression({self.expression} = Array{np.shape(self._value)}"
+        else:
+            return (
+                f"Expression({self.expression} = "
+                f"{str(self._value)[:50]} ... {str(self._value)[-50:]}"
+            )
+
+    # Utility functions for Fourier analysis
+    @staticmethod
+    def fft(s):
+        """Calculate the fast Fourier transform of a signal"""
+        return np.fft.fft(s, norm="forward")
+
+    @staticmethod
+    def ifft(s):
+        """Calculate the inverse fast Fourier transform of a signal"""
+        return np.fft.ifft(s, norm="forward")
+
+    @staticmethod
+    def _find_carrier(s, dt, f0=0.0, bw=np.inf):
+        """Characterize the carrier wave of a signal"""
+        fft_s = np.fft.fft(s, norm="forward")
+        f = np.fft.fftfreq(len(s), d=dt)
+
+        in_band = (f > f0 - 0.5 * bw) & (f < f0 + 0.5 * bw)
+
+        filt_fft_s = np.abs(in_band * fft_s)
+
+        fc = np.abs(f[filt_fft_s == np.max(filt_fft_s)][0])
+        Ac = np.abs(filt_fft_s[filt_fft_s == np.max(filt_fft_s)][0])
+
+        return fc, Ac
+
+    @staticmethod
+    def carrier(s, dt, f0=0.0, bw=np.inf):
+        """Find the carrier frequency of a signal"""
+        return Expression._find_carrier(s, dt, f0=f0, bw=bw)[0]
+
+    @staticmethod
+    def ampl(s, dt, f0=0.0, bw=np.inf):
+        """Calculate the amplitude of the carrier wave in a signal"""
+        return 2 * Expression._find_carrier(s, dt, f0=f0, bw=bw)[1]
+
+    @staticmethod
+    def demod(s, dt, f0, bw=np.inf, cycles=np.inf, filt=None):
+        """Demodulate an oscillatory signal"""
+
+        if np.isfinite(cycles):
+            # Partition signal into segments containing integer number of carrier cycles
+            fc = Expression.carrier(s, dt, f0, bw=bw)  # get carrier frequency
+
+            k_partition = int(cycles / (dt * fc))  # number of samples per partition
+
+            n_parts = int(len(s) / k_partition) + 1
+
+            s_padded = np.concatenate([s, np.zeros(n_parts * k_partition - len(s))])
+
+            partitions = np.reshape(s_padded, (n_parts, k_partition))
+        else:
+            # Demodulate the whole signal
+
+            n_parts = 1
+            k_partition = len(s)
+
+            partitions = np.array([s])
+
+        partitions_demod = np.empty_like(partitions, dtype=np.complex128)
+
+        for i, partition in enumerate(partitions):
+            # Calculate FFT
+            freq = np.fft.fftfreq(k_partition, d=dt)  # frequency values for the FFT
+            fft = np.fft.fft(partition)
+
+            # Find principal frequency within the given band about the given frequency
+            fft_in_positive_band = np.abs(
+                fft * ((freq > f0 - 0.5 * bw) & (freq < f0 + 0.5 * bw))
+            )
+
+            where_f0_closest = np.argwhere(
+                fft_in_positive_band == np.max(fft_in_positive_band)
+            ).flatten()[0]
+
+            if np.abs(fft[where_f0_closest]) > 0.0:
+                phase = np.log(fft[where_f0_closest]).imag  # Get phase of sinusoid
+            else:
+                phase = 0.0
+
+            # Construct the demodulated FFT
+            fft_demod = np.zeros_like(fft)
+
+            # roll the positive component towards zero and remove phase
+            fft_demod += np.roll(fft, -where_f0_closest) * np.exp(-1j * phase)
+
+            # roll the negative component towards zero and remove phase
+            fft_demod += np.roll(fft, where_f0_closest) * np.exp(1j * phase)
+
+            # Apply low pass filter
+            if filt == "gaussian":
+                fft_demod *= np.exp(-(freq**2) / (2 * bw**2))
+            if filt == "sinc":
+                fft_demod *= np.sinc(freq / bw)
+            else:
+                fft_demod *= np.abs(freq) < 0.5 * bw
+
+            partitions_demod[i] = np.fft.ifft(fft_demod)
+
+        signal_demod = partitions_demod.flatten()
+
+        return np.abs(signal_demod)
 
 
 class Remote(Variable):
@@ -308,8 +567,8 @@ class Remote(Variable):
     the server. For Modbus servers, this is the starting address of the
     holding or input register for a read/write or readonly variable,
     respectively. For either set of registers, the starting address is 5*(n-1)
-    for the nth variable in the `readwrite` or `readonly` list/dictionary of
-    variables in the server routine definition.
+    for the nth variable in the `knobs` or `meters` of variables in the server routine
+    definition.
 
     The (optional) `protocol` argument indicates which kind of server to connect
     to. Setting `protocol='modbus'` indicates a Modbus server (controlled by a
@@ -319,10 +578,14 @@ class Remote(Variable):
 
     The `settable` argument is required for remote variables on a Modbus server,
     and is not used for a socket server. If `settable` is set to True, then
-    the variable value is read from the holding registers (`readwrite`
-    variables), otherwise the variable value is read from the input registers
-    (`readonly` variables).
+    the variable value is read from the holding registers (`knobs`), otherwise the
+    variable value is read from the input registers (`meters`).
 
+    The optional `multiplier` and `offset` keyword arguments provide a means to affect
+    a linear transformation of the raw variable value. Readings from the instrument will
+    be multiplied by the `multiplier` and then increased by the `offset`. Set commands
+    to the server will take the variable value, subtract the `offset` and divide by the
+    `multiplier`.
     """
 
     type_map = {
@@ -335,16 +598,24 @@ class Remote(Variable):
     def __init__(
         self,
         server: str,
-        alias: Union[int, str],
+        alias: typing.Union[int, str],
         protocol: str = None,
         settable: bool = False,  # needed for modbus protocol
+        lower_limit: typing.Union[float, int] = None,
+        upper_limit: typing.Union[float, int] = None,
+        multiplier: typing.Union[float, int] = 1,
+        offset: typing.Union[float, int] = 0,
     ):
         self.server = server
         self.alias = alias
         self.protocol = protocol
+        self.lower_limit = lower_limit
+        self.upper_limit = upper_limit
+        self.multiplier = multiplier
+        self.offset = offset
 
         if protocol == "modbus":
-            self._client = instruments.ModbusClient(server)
+            self._client = ModbusClient(server)
             self._settable = settable
 
         else:
@@ -365,27 +636,36 @@ class Remote(Variable):
         """
 
         if self._type is None:
-            self._get_type()
+            self.get_type()
 
         if self.protocol == "modbus":
             fcode = 3 if self.settable else 4
 
-            type_int = self._client.read(fcode, self.alias + 4, _type="16bit_int")
+            if self._type is not None:
 
-            _type = {
-                0: Boolean,
-                1: Toggle,
-                2: Integer,
-                3: Float,
-            }.get(type_int, None)
+                logger.debug(
+                    f'Retrieving value of type {self._type} '
+                    f'starting at register {self.alias}'
+                    f'from Modbus server at {self.server}...'
+                )
 
-            if _type is not None:
                 self._value = self._client.read(
-                    fcode, self.alias, count=4, _type=self._type_map[_type]
+                    fcode, self.alias, count=4, _type=self.type_map[self._type]
+                )
+
+                logger.debug(
+                    f'Value retrieved starting at register {self.alias} '
+                    f'from Modbus server at {self.server} is {self._value}'
                 )
 
         else:
             write_to_socket(self._socket, f"{self.alias} ?")
+
+            logger.debug(
+                f'Retrieving value of type {self._type} '
+                f'with alias {self.alias}'
+                f'from socket server at {self.server}...'
+            )
 
             response = read_from_socket(self._socket, timeout=60, decode=False)
 
@@ -395,23 +675,30 @@ class Remote(Variable):
                 elif b"Error" in response:
                     raise RuntimeError(response.decode().split("Error: ")[-1])
                 else:
+                    bytes_value = response.split(self.alias.encode() + b" ")[-1].strip()
 
-                    bytes_value = response.split(self.alias.encode() + b' ')[-1].strip()
-
-                    if bytes_value[:5] == b'dlpkl':
+                    if bytes_value[:5] == b"dlpkl":
                         # pickled quantity, usually an array, list or tuple
                         self._value = dill.loads(bytes_value[5:])
                     else:
                         self._value = recast(
                             bytes_value,
-                            to=self._type if self._type is not None else Type
+                            to=self._type if self._type is not None else Type,
                         )
 
-            except BaseException as error:
-                print(
-                    f"Warning: unable to retrieve value of {self.alias} "
-                    f'from server at {self.server}; got error "{error}"'
+                logger.debug(
+                    f'Value with alias {self.alias} retrieved '
+                    f'from socket server at {self.server} is {self._value}'
                 )
+
+            except Exception as error:
+                logger.warning(
+                    f"Unable to retrieve value of {self.alias} "
+                    f'from {self.protocol} server at {self.server}: "{error}"'
+                )
+
+        if isinstance(self._value, numbers.Number):
+            self._value = self.multiplier * self._value + self.offset
 
         return self._value
 
@@ -422,22 +709,39 @@ class Remote(Variable):
         Set the value of a remote variable
         """
 
+        if isinstance(value, np.integer):
+            value = (value - int(self.offset)) // int(self.multiplier)
+        elif isinstance(value, np.floating):
+            value = (value - self.offset) / self.multiplier
+
         if self.protocol == "modbus":
-            self._client.write(16, self.alias, value, _type=self._type_map[self._type])
+
+            logger.info(
+                f'Writing value {value} to variable starting at register {self.alias}'
+                f'on Modbus server at {self.server}...'
+            )
+
+            self._client.write(16, self.alias, value, _type=self.type_map[self._type])
 
         else:
+
+            logger.debug(
+                f'Writing value {value} to variable with alias {self.alias} '
+                f'on socket server at {self.server}...'
+            )
+
             write_to_socket(self._socket, f"{self.alias} {value}")
 
             check = read_from_socket(self._socket, timeout=60)
 
             if check == "" or check is None:
-                print(
-                    f"Warning: received no response from server at "
+                logger.warning(
+                    f"Received no response from server at "
                     f"{self.server} while trying to set {self.alias}"
                 )
             elif "Error" in check:
-                print(
-                    f'Warning: got response "{check}" while trying to set '
+                logger.warning(
+                    f'Got response "{check}" while trying to set '
                     f"{self.alias} on server at {self.server}"
                 )
             else:
@@ -445,21 +749,21 @@ class Remote(Variable):
                     check_value = recast(check.split(f"{self.alias} ")[1])
 
                     if value != check_value:
-                        print(
-                            f"Warning: attempted to set {self.alias} on "
+                        logger.warning(
+                            f"Attempted to set {self.alias} on "
                             f"server at {self.server} to {value} but "
                             f"checked value is {check_value}"
                         )
 
                 except ValueError as val_err:
-                    print(
-                        f"Warning: unable to check value while setting "
+                    logger.warning(
+                        f"Unable to check value while setting "
                         f"{self.alias} on server at {self.server}; "
                         f'got error "{val_err}"'
                     )
                 except IndexError as ind_err:
-                    print(
-                        f"Warning: unable to check value while setting "
+                    logger.warning(
+                        f"Unable to check value while setting "
                         f"{self.alias} on server at {self.server}; "
                         f'got error "{ind_err}"'
                     )
@@ -473,16 +777,52 @@ class Remote(Variable):
 
     def get_type(self):
         """Get the data type of the remote variable"""
-        write_to_socket(self._socket, f"{self.alias} type?")
 
-        response = read_from_socket(self._socket, timeout=60)
+        if self.protocol == "modbus":
 
-        if response is not None:
-            for _type in types.supported:
-                if str(_type) in response.split(self.alias)[-1]:
-                    self._type = types.supported.get(_type, None)
+            logger.debug(
+                f'Getting data type of variable starting at register {self.alias}'
+                f'on Modbus server at {self.server}...'
+            )
+
+            fcode = 3 if self.settable else 4
+
+            type_int = self._client.read(fcode, self.alias + 4, _type="16bit_int")
+
+            self._type = {
+                0: Boolean,
+                1: Toggle,
+                2: Integer,
+                3: Float,
+            }.get(type_int, None)
+
+            logger.debug(
+                f'Data type of variable starting at register {self.alias} '
+                f'on Modbus server at {self.server} is {self._type}'
+            )
+
         else:
-            self._type = None
+
+            logger.debug(
+                f'Getting data type of variable with alias {self.alias}'
+                f'on socket server at {self.server}...'
+            )
+
+            write_to_socket(self._socket, f"{self.alias} type?")
+
+            response = read_from_socket(self._socket, timeout=60)
+
+            if response is not None:
+                for _type in supported_types:
+                    if str(_type) in response.split(self.alias)[-1]:
+                        self._type = supported_types.get(_type, None)
+            else:
+                self._type = None
+
+            logger.debug(
+                f'Data type of variable with alias {self.alias} '
+                f'on socket server at {self.server} is {self._type}'
+            )
 
     def get_settable(self):
         """Get settability of remote variable"""
@@ -490,6 +830,9 @@ class Remote(Variable):
 
         response = read_from_socket(self._socket, timeout=60)
         self._settable = response == f"{self.alias} settable"
+
+    def __str__(self):
+        return f"Remote({self.alias}@{self.server} = {self.value})"
 
 
 class Parameter(Variable):
@@ -503,13 +846,19 @@ class Parameter(Variable):
 
     _settable = True  #:
 
-    def __init__(self, parameter: Type):
-        self.parameter = parameter
-        self._value = parameter
+    def __init__(
+            self, parameter: typing.Union[float, int, bool, str, Toggle, np.ndarray]
+    ):
+        self._value = recast(parameter)
 
-        for name, _type in types.supported.items():
-            if isinstance(parameter, _type):
+        for name, _type in supported_types.items():
+            if isinstance(self._value, _type) and _type is not Type:
+
                 self._type = _type
+
+        logger.debug(
+            f'Setting data type of parameter {self._value} to {self._type}'
+        )
 
     @property
     @Variable.getter_type_validator
@@ -521,8 +870,10 @@ class Parameter(Variable):
     @Variable.setter_type_validator
     def value(self, value):
         """Set the parameter value"""
-        self.parameter = value
         self._value = value
+
+    def __str__(self):
+        return f"Parameter({self.value})"
 
 
 supported = {
